setup-requirements.txt
sanic
sanic-jinja2
aioredis
beautifulsoup4
PyYAML
lxml
aiohttp
asyncpg